refactor: splitting the extract script into testable methods
