+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": null,
     "lines": null
   },
-  "generated_at": "2021-07-30T17:41:52Z",
+  "generated_at": "2021-08-17T22:36:25Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -42,14 +42,6 @@
     }
   ],
   "results": {
-    "eccc/extract/app/extract_zips.py": [
-      {
-        "hashed_secret": "e41dab5d3a405c79a9e1c8fcf63b7292eab80325",
-        "is_verified": false,
-        "line_number": 7,
-        "type": "Hex High Entropy String"
-      }
-    ],
     "ggircs-app/app/.env": [
       {
         "hashed_secret": "5bf1fd927dfb8679496a2e6cf00cbe50c1c87145",
